added standalone query file to common
