+++ common/amundsen_common/models/feature.py
@@ -13,19 +13,6 @@ from marshmallow3_annotations.ext.attrs import AttrsSchema
 
 
 @attr.s(auto_attribs=True, kw_only=True)
-class Query:
-    name: Optional[str]
-    text: str
-    url: Optional[str]
-
-
-class QuerySchema(AttrsSchema):
-    class Meta:
-        target = Query
-        register_as_scheme = True
-
-
-@attr.s(auto_attribs=True, kw_only=True)
 class ColumnItem:
     column_name: str
     column_type: str
